requirements.txt
# 公文渲染引擎 V14 - Python 依赖

# 核心依赖
python-docx>=0.8.11

# 类型提示
typing-extensions>=4.8.0

# 数据校验
pydantic>=2.0.0
